archive/custom_GPT/requirements.txt
python_dotenv
wikipedia
tiktoken
neo4j
streamlit
Pillow
fastapi
PyPDF2
pydantic
uvicorn
sse_starlette
boto3
streamlit
# missing from the langchain base image?
langchain_openai
langchain_community
langchain_google_genai
langchain_ollama
langchain_huggingface
langchain_aws
langchain_neo4j